+++ sdk/keyvault/Azure.Security.KeyVault.Keys/src/Cryptography/ICryptographyProvider.cs
@@ -12,13 +12,13 @@ namespace Azure.Security.KeyVault.Keys.Cryptography
 
         bool SupportsOperation(KeyOperation operation);
 
-        Task<EncryptResult> EncryptAsync(EncryptionAlgorithm algorithm, byte[] plaintext, CancellationToken cancellationToken = default);
+        Task<EncryptResult> EncryptAsync(EncryptionAlgorithm algorithm, EncryptOptions options, CancellationToken cancellationToken = default);
 
-        EncryptResult Encrypt(EncryptionAlgorithm algorithm, byte[] plaintext, CancellationToken cancellationToken = default);
+        EncryptResult Encrypt(EncryptionAlgorithm algorithm, EncryptOptions options, CancellationToken cancellationToken = default);
 
-        Task<DecryptResult> DecryptAsync(EncryptionAlgorithm algorithm, byte[] ciphertext, CancellationToken cancellationToken = default);
+        Task<DecryptResult> DecryptAsync(EncryptionAlgorithm algorithm, DecryptOptions options, CancellationToken cancellationToken = default);
 
-        DecryptResult Decrypt(EncryptionAlgorithm algorithm, byte[] ciphertext, CancellationToken cancellationToken = default);
+        DecryptResult Decrypt(EncryptionAlgorithm algorithm, DecryptOptions options, CancellationToken cancellationToken = default);
 
         Task<WrapResult> WrapKeyAsync(KeyWrapAlgorithm algorithm, byte[] key, CancellationToken cancellationToken = default);