+++ sdk/keyvault/Azure.Security.KeyVault.Keys/src/Cryptography/RemoteCryptographyClient.cs
@@ -41,21 +41,17 @@ namespace Azure.Security.KeyVault.Keys.Cryptography
 
         public bool SupportsOperation(KeyOperation operation) => true;
 
-        public virtual async Task<Response<EncryptResult>> EncryptAsync(EncryptionAlgorithm algorithm, byte[] plaintext, CancellationToken cancellationToken = default)
+        public virtual async Task<Response<EncryptResult>> EncryptAsync(EncryptionAlgorithm algorithm, EncryptOptions options, CancellationToken cancellationToken = default)
         {
-            var parameters = new KeyEncryptParameters()
-            {
-                Algorithm = algorithm.ToString(),
-                Value = plaintext,
-            };
-
             using DiagnosticScope scope = Pipeline.CreateScope($"{nameof(RemoteCryptographyClient)}.{nameof(Encrypt)}");
             scope.AddAttribute("key", _keyId);
             scope.Start();
 
             try
             {
-                return await Pipeline.SendRequestAsync(RequestMethod.Post, parameters, () => new EncryptResult { Algorithm = algorithm }, cancellationToken, "/encrypt").ConfigureAwait(false);
+                options.Initialize(algorithm);
+
+                return await Pipeline.SendRequestAsync(RequestMethod.Post, options, () => new EncryptResult { Algorithm = algorithm }, cancellationToken, "/encrypt").ConfigureAwait(false);
             }
             catch (Exception e)
             {
@@ -64,21 +60,17 @@ namespace Azure.Security.KeyVault.Keys.Cryptography
             }
         }
 
-        public virtual Response<EncryptResult> Encrypt(EncryptionAlgorithm algorithm, byte[] plaintext, CancellationToken cancellationToken = default)
+        public virtual Response<EncryptResult> Encrypt(EncryptionAlgorithm algorithm, EncryptOptions options, CancellationToken cancellationToken = default)
         {
-            var parameters = new KeyEncryptParameters()
-            {
-                Algorithm = algorithm.ToString(),
-                Value = plaintext,
-            };
-
             using DiagnosticScope scope = Pipeline.CreateScope($"{nameof(RemoteCryptographyClient)}.{nameof(Encrypt)}");
             scope.AddAttribute("key", _keyId);
             scope.Start();
 
             try
             {
-                return Pipeline.SendRequest(RequestMethod.Post, parameters, () => new EncryptResult { Algorithm = algorithm }, cancellationToken, "/encrypt");
+                options.Initialize(algorithm);
+
+                return Pipeline.SendRequest(RequestMethod.Post, options, () => new EncryptResult { Algorithm = algorithm }, cancellationToken, "/encrypt");
             }
             catch (Exception e)
             {
@@ -87,21 +79,17 @@ namespace Azure.Security.KeyVault.Keys.Cryptography
             }
         }
 
-        public virtual async Task<Response<DecryptResult>> DecryptAsync(EncryptionAlgorithm algorithm, byte[] ciphertext, CancellationToken cancellationToken = default)
+        public virtual async Task<Response<DecryptResult>> DecryptAsync(EncryptionAlgorithm algorithm, DecryptOptions options, CancellationToken cancellationToken = default)
         {
-            var parameters = new KeyEncryptParameters()
-            {
-                Algorithm = algorithm.ToString(),
-                Value = ciphertext,
-            };
-
             using DiagnosticScope scope = Pipeline.CreateScope($"{nameof(RemoteCryptographyClient)}.{nameof(Decrypt)}");
             scope.AddAttribute("key", _keyId);
             scope.Start();
 
             try
             {
-                return await Pipeline.SendRequestAsync(RequestMethod.Post, parameters, () => new DecryptResult { Algorithm = algorithm }, cancellationToken, "/decrypt").ConfigureAwait(false);
+                options.Algorithm = algorithm;
+
+                return await Pipeline.SendRequestAsync(RequestMethod.Post, options, () => new DecryptResult { Algorithm = algorithm }, cancellationToken, "/decrypt").ConfigureAwait(false);
             }
             catch (Exception e)
             {
@@ -110,21 +98,17 @@ namespace Azure.Security.KeyVault.Keys.Cryptography
             }
         }
 
-        public virtual Response<DecryptResult> Decrypt(EncryptionAlgorithm algorithm, byte[] ciphertext, CancellationToken cancellationToken = default)
+        public virtual Response<DecryptResult> Decrypt(EncryptionAlgorithm algorithm, DecryptOptions options, CancellationToken cancellationToken = default)
         {
-            var parameters = new KeyEncryptParameters()
-            {
-                Algorithm = algorithm.ToString(),
-                Value = ciphertext,
-            };
-
             using DiagnosticScope scope = Pipeline.CreateScope($"{nameof(RemoteCryptographyClient)}.{nameof(Decrypt)}");
             scope.AddAttribute("key", _keyId);
             scope.Start();
 
             try
             {
-                return Pipeline.SendRequest(RequestMethod.Post, parameters, () => new DecryptResult { Algorithm = algorithm }, cancellationToken, "/decrypt");
+                options.Algorithm = algorithm;
+
+                return Pipeline.SendRequest(RequestMethod.Post, options, () => new DecryptResult { Algorithm = algorithm }, cancellationToken, "/decrypt");
             }
             catch (Exception e)
             {
@@ -355,24 +339,24 @@ namespace Azure.Security.KeyVault.Keys.Cryptography
 
         bool ICryptographyProvider.ShouldRemote => false;
 
-        async Task<EncryptResult> ICryptographyProvider.EncryptAsync(EncryptionAlgorithm algorithm, byte[] plaintext, CancellationToken cancellationToken)
+        async Task<EncryptResult> ICryptographyProvider.EncryptAsync(EncryptionAlgorithm algorithm, EncryptOptions options, CancellationToken cancellationToken)
         {
-            return await EncryptAsync(algorithm, plaintext, cancellationToken).ConfigureAwait(false);
+            return await EncryptAsync(algorithm, options, cancellationToken).ConfigureAwait(false);
         }
 
-        EncryptResult ICryptographyProvider.Encrypt(EncryptionAlgorithm algorithm, byte[] plaintext, CancellationToken cancellationToken)
+        EncryptResult ICryptographyProvider.Encrypt(EncryptionAlgorithm algorithm, EncryptOptions options, CancellationToken cancellationToken)
         {
-            return Encrypt(algorithm, plaintext, cancellationToken);
+            return Encrypt(algorithm, options, cancellationToken);
         }
 
-        async Task<DecryptResult> ICryptographyProvider.DecryptAsync(EncryptionAlgorithm algorithm, byte[] ciphertext, CancellationToken cancellationToken)
+        async Task<DecryptResult> ICryptographyProvider.DecryptAsync(EncryptionAlgorithm algorithm, DecryptOptions options, CancellationToken cancellationToken)
         {
-            return await DecryptAsync(algorithm, ciphertext, cancellationToken).ConfigureAwait(false);
+            return await DecryptAsync(algorithm, options, cancellationToken).ConfigureAwait(false);
         }
 
-        DecryptResult ICryptographyProvider.Decrypt(EncryptionAlgorithm algorithm, byte[] ciphertext, CancellationToken cancellationToken)
+        DecryptResult ICryptographyProvider.Decrypt(EncryptionAlgorithm algorithm, DecryptOptions options, CancellationToken cancellationToken)
         {
-            return Decrypt(algorithm, ciphertext, cancellationToken);
+            return Decrypt(algorithm, options, cancellationToken);
         }
 
         async Task<WrapResult> ICryptographyProvider.WrapKeyAsync(KeyWrapAlgorithm algorithm, byte[] key, CancellationToken cancellationToken)