+++ sdk/keyvault/Azure.Security.KeyVault.Keys/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Added `KeyType.OctHsm` to support "oct-HSM" key operations.
+- Added AES-GCM and AES-CBC support for encrypting and decrypting.
 
 ## 4.2.0-beta.2 (2020-10-06)